Ensure the base shapes have alpha channels from the start
Some of the base shapes didn't have alpha channels, and so they still
had a white background around some of the shapes when being placed on.
Instead now, at the beginning we'll make sure that it has an alpha
channel, so that the generation will work with or without it from the
beginning.