Build full model only on tags
Doing one for each master commit is a bit extra. Especially considering
that tags are also run along side another master commit, leading to
double builds. If we need to check the accuracy of a build in the middle
of a pull request, we can just push a dev tag for the next version.